split qlinear fully connected family
