+++ onnx2tf/tflite_builder/op_builders/__init__.py
@@ -173,14 +173,12 @@ from onnx2tf.tflite_builder.op_builders.control import (
 from onnx2tf.tflite_builder.op_builders.quantized import (
     build_conv_integer_op,
     build_dequantize_linear_op,
-    build_qgemm_op,
     build_qlinear_add_op,
     build_qlinear_average_pool_op,
     build_qlinear_concat_op,
     build_qlinear_conv_op,
     build_qlinear_global_average_pool_op,
     build_qlinear_leaky_relu_op,
-    build_qlinear_matmul_op,
     build_qlinear_mul_op,
     build_qlinear_sigmoid_op,
     build_qlinear_softmax_op,
@@ -188,6 +186,10 @@ from onnx2tf.tflite_builder.op_builders.quantized import (
 )
 from onnx2tf.tflite_builder.op_builders.dynamic_quantize import (
     build_dynamic_quantize_linear_op,
+)
+from onnx2tf.tflite_builder.op_builders.qlinear_fc import (
+    build_qgemm_op,
+    build_qlinear_matmul_op,
 )
 
 __all__ = [
